pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "copper-mining-risk-model"
version = "1.0.0"
description = "Mining planning and performance analytics pipeline with BI-ready outputs and advanced valuation and risk modules."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "openpyxl>=3.1",
  "pandas>=2.2",
  "pyyaml>=6.0",
  "scipy>=1.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
